requirements.txt
discord.py[voice]>=1.5.0
colorama>=0.0.0
pydub>=0.0.0
regex
fuzzywuzzy